Refactor the `mock` impl into a separate module.
This is intended to more clearly separate the interface exposed to
clients from the mock impl.
